split StandardSecurityHandler into two smaller classes
