+++ lib/pdf/reader.rb
@@ -316,7 +316,8 @@ require 'pdf/reader/register_receiver'
 require 'pdf/reader/null_security_handler'
 require 'pdf/reader/security_handler_factory'
 require 'pdf/reader/standard_key_builder'
-require 'pdf/reader/standard_security_handler'
+require 'pdf/reader/aes_v2_security_handler'
+require 'pdf/reader/rc4_security_handler'
 require 'pdf/reader/standard_security_handler_v5'
 require 'pdf/reader/unimplemented_security_handler'
 require 'pdf/reader/stream'
